[add] third packages

+++ entry.sh
@@ -50,10 +50,14 @@ mkdir -p /etc/kubecube/down
 mkdir -p /etc/kubecube/bin
 cd /etc/kubecube
 
-echo -e "$(date +'%Y-%m-%d %H:%M:%S') \033[32mINFO\033[0m downloading manifests for kubecube"
-wget https://gitee.com/kubecube/manifests/repository/archive/master.zip -O manifests.zip
+if [ -e "manifests.zip" ]; then
+  echo -e "$(date +'%Y-%m-%d %H:%M:%S') \033[32mINFO\033[0m manifests already exist"
+else
+  echo -e "$(date +'%Y-%m-%d %H:%M:%S') \033[32mINFO\033[0m downloading manifests for kubecube"
+  wget https://gitee.com/kubecube/manifests/repository/archive/master.zip -O manifests.zip
 
-unzip manifests.zip > /dev/null
+  unzip manifests.zip > /dev/null
+fi
 
 if [[ ${CUSTOMIZE} = "true" ]]; then
   echo -e "\033[32m================================================\033[0m"
